requirements.txt
fastapi
openai
scikit-learn
bertopic
sentence-transformers
umap-learn
numpy
transformers
pandas
datasets
torch